+++ src/main/java/net/chaos/chaosmod/items/tools/ToolHoe.java
@@ -1,23 +1,10 @@
 package net.chaos.chaosmod.items.tools;
 
-import java.util.List;
-
-import net.chaos.chaosmod.Main;
 import net.chaos.chaosmod.init.ModItems;
 import net.chaos.chaosmod.tabs.ModTabs;
-import net.minecraft.client.gui.GuiScreen;
-import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemHoe;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.text.TextComponentString;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.World;
 import proxy.IItemModel;
-import util.Reference;
 
 public class ToolHoe extends ItemHoe implements IItemModel {
 
@@ -27,24 +14,6 @@ public class ToolHoe extends ItemHoe implements IItemModel {
 		setRegistryName(name);
 		
 		ModItems.ITEMS.add(this);
-	}
-	
-	@Override
-	public void addInformation(ItemStack stack, World worldIn, List<String> tooltip, ITooltipFlag flagIn) {
-		if (GuiScreen.isShiftKeyDown()) {
-			tooltip.add(new TextComponentString(TextFormatting.DARK_RED + "Shift+Right-click to launch credits").getFormattedText());
-		} else {
-			tooltip.add(new TextComponentString("Press " + TextFormatting.YELLOW + "[SHIFT]" + TextFormatting.RESET + " to see further actions").getFormattedText());
-		}
-		super.addInformation(stack, worldIn, tooltip, flagIn);
-	}
-	
-	@Override
-	public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
-		if (GuiScreen.isShiftKeyDown()) {
-			playerIn.openGui(Main.instance, Reference.GUI_CREDITS_ID, worldIn, 0, 0, 0);
-		}
-		return super.onItemRightClick(worldIn, playerIn, handIn);
 	}
 
 	@Override